split plugins in export / import files
